Completed all functions exercises

+++ exercises/functions/functions4.rs
@@ -14,7 +14,7 @@ fn main() {
     println!("Your sale price is {}", sale_price(original_price));
 }
 
-fn sale_price(price: i32) -> i32 {
+fn sale_price(price: i8) -> i8 {
     if is_even(price) {
         price - 10
     } else {
@@ -22,6 +22,6 @@ fn sale_price(price: i32) -> i32 {
     }
 }
 
-fn is_even(num: i32) -> bool {
+fn is_even(num: i8) -> bool {
     num % 2 == 0
 }
